[Fix]ログイン後にsession[:return_to]がない場合に対応、デフォルトパス再設定

+++ app/controllers/user_sessions_controller.rb
@@ -4,8 +4,12 @@ class UserSessionsController < ApplicationController
   def create
     @user = login(params[:email], params[:password], params[:remember])
     if @user
-      redirect_back_or_to session[:return_to], success: "ログインしました"
-      session.delete(:return_to)
+      if session[:return_to]
+        redirect_to session[:return_to], success: "ログインしました"
+        session.delete(:return_to)
+      else
+        redirect_back_or_to diaries_url, success: "ログインしました"
+      end
     else
       flash.now[:danger] = "ログインに失敗しました"
       render :new
